19.02 commit jdbc classes added

+++ src/main/java/com/ryszard/dao/EmployeeDao.java
@@ -2,5 +2,5 @@ package com.ryszard.dao;
 
 import com.ryszard.domain.vo.Employee;
 
-public interface EmployeeDao extends GenericDao<Employee, Long> {
+public interface EmployeeDao extends GenericDao<Employee, Integer> {
 }
